Mze emoji
When you run this code, it will generate a random maze and display it on the page using HTML and CSS. The emoji will be represented by a smiling face (😊) and will start from a yellow cell and attempt to reach the green cell in the center (🏁). However, please note that this example does not include the maze-solving logic or dynamic adaptation you mentioned. Implementing such complex functionality would require additional programming languages like JavaScript or Python